ODPM-419 moved cassandra specific utils under cassandra/query.go
